+++ Libplanet.Net.Tests/SwarmTest.Preload.cs
@@ -377,7 +377,7 @@ namespace Libplanet.Net.Tests
         public async Task NoRenderInPreload()
         {
             var policy = new BlockPolicy<DumbAction>(new MinerReward(1));
-            var renderer = new RecordingActionRenderer<DumbAction>();
+            var renderer = new RecordingActionRenderer();
             var chain = MakeBlockChain(
                 policy,
                 new MemoryStore(),