+++ Libplanet/Blockchain/Renderers/AnonymousRenderer.cs
@@ -1,5 +1,4 @@
 using System;
-using Libplanet.Action;
 using Libplanet.Blocks;
 
 namespace Libplanet.Blockchain.Renderers
@@ -7,7 +6,7 @@ namespace Libplanet.Blockchain.Renderers
     /// <summary>
     /// A renderer that invokes its callbacks.
     /// <para>This class is useful when you want an one-use ad-hoc implementation (i.e., Java-style
-    /// anonymous class) of <see cref="IRenderer{T}"/> interface.</para>
+    /// anonymous class) of <see cref="IRenderer"/> interface.</para>
     /// </summary>
     /// <example>
     /// With object initializers, you can easily make an one-use renderer:
@@ -21,10 +20,7 @@ namespace Libplanet.Blockchain.Renderers
     /// };
     /// </code>
     /// </example>
-    /// <typeparam name="T">An <see cref="IAction"/> type.  It should match to
-    /// <see cref="BlockChain{T}"/>'s type parameter.</typeparam>
-    public class AnonymousRenderer<T> : IRenderer<T>
-        where T : IAction, new()
+    public class AnonymousRenderer : IRenderer
     {
         /// <summary>
         /// A callback function to be invoked together with
@@ -32,7 +28,7 @@ namespace Libplanet.Blockchain.Renderers
         /// </summary>
         public Action<Block, Block>? BlockRenderer { get; set; }
 
-        /// <inheritdoc cref="IRenderer{T}.RenderBlock(Block, Block)"/>
+        /// <inheritdoc cref="IRenderer.RenderBlock(Block, Block)"/>
         public void RenderBlock(Block oldTip, Block newTip) =>
             BlockRenderer?.Invoke(oldTip, newTip);
     }