+++ Libplanet/Blockchain/Renderers/AtomicActionRenderer.cs
@@ -8,31 +8,28 @@ namespace Libplanet.Blockchain.Renderers
 {
     /// <summary>
     /// A middleware to make action render events to satisfy transactions' atomicity.
-    /// <para>Decorates an <see cref="IActionRenderer{T}"/> instance and filters out render events
+    /// <para>Decorates an <see cref="IActionRenderer"/> instance and filters out render events
     /// made by unsuccessful transactions (i.e., transactions with one or more exception-throwing
     /// actions).</para>
     /// </summary>
     /// <remarks>The wrapped <see cref="ActionRenderer"/> will not receive any
-    /// <see cref="IActionRenderer{T}.RenderActionError"/> events except for block actions,
+    /// <see cref="IActionRenderer.RenderActionError"/> events except for block actions,
     /// which do not belong to any transactions.
     /// </remarks>
-    /// <typeparam name="T">An <see cref="IAction"/> type.  It should match to
-    /// <see cref="BlockChain{T}"/>'s type parameter.</typeparam>
-    public sealed class AtomicActionRenderer<T> : IActionRenderer<T>
-        where T : IAction, new()
+    public sealed class AtomicActionRenderer : IActionRenderer
     {
         private readonly List<(IAction, IActionContext, IAccountStateDelta)> _eventBuffer;
         private TxId? _lastTxId;
         private bool _errored;
 
         /// <summary>
-        /// Creates a new <see cref="AtomicActionRenderer{T}"/> instance decorating the given
+        /// Creates a new <see cref="AtomicActionRenderer"/> instance decorating the given
         /// <paramref name="actionRenderer"/>.
         /// </summary>
         /// <param name="actionRenderer">The inner action renderer which has the <em>actual</em>
         /// implementations and expects to receive no <see cref="RenderActionError"/> events.
         /// </param>
-        public AtomicActionRenderer(IActionRenderer<T> actionRenderer)
+        public AtomicActionRenderer(IActionRenderer actionRenderer)
         {
             ActionRenderer = actionRenderer;
             _lastTxId = null;
@@ -44,15 +41,15 @@ namespace Libplanet.Blockchain.Renderers
         /// The inner action renderer which has the <em>actual</em> implementations and expects to
         /// receive no <see cref="RenderActionError"/> events.
         /// </summary>
-        public IActionRenderer<T> ActionRenderer { get; }
+        public IActionRenderer ActionRenderer { get; }
 
-        /// <inheritdoc cref="IRenderer{T}.RenderBlock(Block, Block)"/>
+        /// <inheritdoc cref="IRenderer.RenderBlock(Block, Block)"/>
         public void RenderBlock(Block oldTip, Block newTip)
         {
             ActionRenderer.RenderBlock(oldTip, newTip);
         }
 
-        /// <inheritdoc cref="IActionRenderer{T}.RenderBlockEnd(Block, Block)"/>
+        /// <inheritdoc cref="IActionRenderer.RenderBlockEnd(Block, Block)"/>
         public void RenderBlockEnd(Block oldTip, Block newTip)
         {
             FlushBuffer(null, ActionRenderer.RenderAction);
@@ -60,7 +57,7 @@ namespace Libplanet.Blockchain.Renderers
         }
 
         /// <inheritdoc
-        /// cref="IActionRenderer{T}.RenderAction(IAction, IActionContext, IAccountStateDelta)"/>
+        /// cref="IActionRenderer.RenderAction(IAction, IActionContext, IAccountStateDelta)"/>
         public void RenderAction(
             IAction action,
             IActionContext context,
@@ -83,7 +80,7 @@ namespace Libplanet.Blockchain.Renderers
         }
 
         /// <inheritdoc
-        /// cref="IActionRenderer{T}.RenderActionError(IAction, IActionContext, Exception)"/>
+        /// cref="IActionRenderer.RenderActionError(IAction, IActionContext, Exception)"/>
         public void RenderActionError(IAction action, IActionContext context, Exception exception)
         {
             if (!context.TxId.Equals(_lastTxId))